Moving search suggestions into their own package
